+++ samples/spring-boot-plain/src/main/java/io/javaoperatorsdk/operator/sample/Config.java
@@ -2,7 +2,7 @@ package io.javaoperatorsdk.operator.sample;
 
 import io.javaoperatorsdk.operator.Operator;
 import io.javaoperatorsdk.operator.api.ResourceController;
-import io.javaoperatorsdk.operator.api.config.DefaultConfigurationService;
+import io.javaoperatorsdk.operator.DefaultConfigurationService;
 import io.fabric8.kubernetes.client.DefaultKubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.javaoperatorsdk.operator.Operator;
